Properly implemented MoveEffect class & Move data
