+++ PMDMMO-common/src/com/darkxell/common/move/MoveEffect.java
@@ -1,6 +1,23 @@
 package com.darkxell.common.move;
 
-public class MoveEffect
+import java.util.ArrayList;
+
+import com.darkxell.common.ai.AIUtils;
+import com.darkxell.common.dungeon.floor.Floor;
+import com.darkxell.common.dungeon.floor.Room;
+import com.darkxell.common.dungeon.floor.Tile;
+import com.darkxell.common.dungeon.floor.TileType;
+import com.darkxell.common.event.DungeonEvent;
+import com.darkxell.common.event.DungeonEvent.MessageEvent;
+import com.darkxell.common.event.move.MoveSelectionEvent.MoveUse;
+import com.darkxell.common.event.move.MoveUseEvent;
+import com.darkxell.common.move.Move.MoveRange;
+import com.darkxell.common.pokemon.DungeonPokemon;
+import com.darkxell.common.pokemon.PokemonType;
+import com.darkxell.common.util.Direction;
+import com.darkxell.common.util.language.Message;
+
+public abstract class MoveEffect
 {
 
 	public final int id;
@@ -10,5 +27,172 @@ public class MoveEffect
 		this.id = id;
 		MoveEffects.effects.put(this.id, this);
 	}
+
+	/** Removes all Pokémon this move is not supposed to target. */
+	protected void filterTargets(Floor floor, Move move, DungeonPokemon user, ArrayList<DungeonPokemon> targets)
+	{
+		switch (move.targets)
+		{
+			case All:
+				break;
+
+			case Allies:
+				targets.removeIf((DungeonPokemon p) -> p == user || !p.isAlliedWith(user));
+				break;
+
+			case Foes:
+				targets.removeIf((DungeonPokemon p) -> p == user || p.isAlliedWith(user));
+				break;
+
+			case Others:
+				targets.remove(user);
+				break;
+
+			case Team:
+				targets.removeIf((DungeonPokemon p) -> !p.isAlliedWith(user));
+				break;
+
+			case User:
+				targets.removeIf((DungeonPokemon p) -> p != user);
+				break;
+
+			case None:
+				targets.clear();
+				break;
+		}
+	}
+
+	/** @param move
+	 * @param user - The Pokémon using this Move.
+	 * @param floor - The Floor context.
+	 * @return The Pokémon affected by this Move. */
+	public DungeonPokemon[] getTargets(Move move, DungeonPokemon user, Floor floor)
+	{
+		ArrayList<DungeonPokemon> targets = new ArrayList<DungeonPokemon>();
+		Tile t = user.tile(), front = t.adjacentTile(user.facing());
+
+		switch (move.range)
+		{
+			case Ambient:
+				targets.add(null);
+				break;
+
+			case Around:
+				for (Direction d : Direction.directions)
+					if (t.adjacentTile(d).getPokemon() != null) targets.add(t.adjacentTile(d).getPokemon());
+				break;
+
+			case Floor:
+				targets.addAll(floor.listPokemon());
+				break;
+
+			case Front_row:
+				for (Direction d : new Direction[] { user.facing().rotateCounterClockwise(), user.facing(), user.facing().rotateClockwise() })
+					if (t.adjacentTile(d).getPokemon() != null) targets.add(t.adjacentTile(d).getPokemon());
+				break;
+
+			case Line:
+				int distance = 0;
+				boolean done;
+				Tile current = t;
+				do
+				{
+					current = current.adjacentTile(user.facing());
+					if (current.getPokemon() != null) targets.add(current.getPokemon());
+					++distance;
+					done = !targets.isEmpty() || distance > 10 || current.type() == TileType.WALL || current.type() == TileType.WALL_END;
+				} while (!done);
+				break;
+
+			case Room:
+				Room r = floor.roomAt(user.tile().x, user.tile().y);
+				if (r == null)
+				{
+					for (Tile tile : AIUtils.visibleTiles(floor, user))
+						if (tile.getPokemon() != null) targets.add(tile.getPokemon());
+				} else for (Tile t2 : r.listTiles())
+					if (t2.getPokemon() != null) targets.add(t2.getPokemon());
+				break;
+
+			case Self:
+				targets.add(user);
+				break;
+
+			case Two_tiles:
+				if (front.getPokemon() != null) targets.add(front.getPokemon());
+				else if (front.type().canWalkOn(user))
+				{
+					Tile behind = front.adjacentTile(user.facing());
+					if (behind.getPokemon() != null) targets.add(behind.getPokemon());
+				}
+				break;
+
+			case Front:
+			case Front_corners:
+			default:
+				DungeonPokemon f = user.tile().adjacentTile(user.facing()).getPokemon();
+				if (f != null)
+				{
+					boolean valid = true;
+					if (user.facing().isDiagonal() && move.range != MoveRange.Front_corners)
+					{
+						Tile t1 = user.tile().adjacentTile(user.facing().rotateClockwise());
+						if (t1.type() == TileType.WALL || t1.type() == TileType.WALL_END) valid = false;
+						t1 = user.tile().adjacentTile(user.facing().rotateCounterClockwise());
+						if (t1.type() == TileType.WALL || t1.type() == TileType.WALL_END) valid = false;
+					}
+					if (valid) targets.add(f);
+				}
+		}
+
+		this.filterTargets(floor, move, user, targets);
+		if (move.range == MoveRange.Room || move.range == MoveRange.Floor)
+			targets.sort((DungeonPokemon p1, DungeonPokemon p2) -> floor.dungeon.compare(p1, p2));
+
+		return targets.toArray(new DungeonPokemon[targets.size()]);
+	}
+
+	/** Main method called when a Pokémon uses a Move on a target. */
+	public void mainUse(MoveUse usedMove, DungeonPokemon target, Floor floor, ArrayList<DungeonEvent> events)
+	{
+		Move move = usedMove.move.move();
+		boolean missed = this.misses(usedMove, target, floor);
+		float effectiveness = move.type == null ? PokemonType.NORMALLY_EFFECTIVE : move.type.effectivenessOn(target.species());
+		if (effectiveness == PokemonType.NO_EFFECT) events.add(new MessageEvent(floor, move.unaffectedMessage(target)));
+		else
+		{
+			if (!missed && this != MoveEffects.Basic_attack) target.receiveMove(usedMove.move.isLinked() ? DungeonPokemon.LINKED_MOVES : DungeonPokemon.MOVES);
+			if (move.power != -1)
+			{
+				if (effectiveness == PokemonType.SUPER_EFFECTIVE)
+					events.add(new MessageEvent(floor, new Message("move.effectiveness.super").addReplacement("<pokemon>", target.getNickname())));
+				else if (effectiveness == PokemonType.NOT_VERY_EFFECTIVE)
+					events.add(new MessageEvent(floor, new Message("move.effectiveness.not_very").addReplacement("<pokemon>", target.getNickname())));
+				this.useOn(usedMove, target, floor, missed, events);
+			}
+		}
+	}
+
+	/** @param usedMove - The Move used.
+	 * @param target - The Pokémon receiving the Move.
+	 * @param floor - The Floor context.
+	 * @return True if this Move misses. */
+	public boolean misses(MoveUse usedMove, DungeonPokemon target, Floor floor)
+	{
+		return false;
+	}
+
+	public void prepareUse(MoveUse move, Floor floor, ArrayList<DungeonEvent> events)
+	{
+		Move m = move.move.move();
+
+		DungeonPokemon[] pokemon = this.getTargets(m, move.user, floor);
+		for (int i = 0; i < pokemon.length; ++i)
+			events.add(new MoveUseEvent(floor, move, pokemon[i]));
+
+		if (events.size() == 0 && this != MoveEffects.Basic_attack) events.add(new MessageEvent(floor, new Message("move.no_target")));
+	}
+
+	protected abstract void useOn(MoveUse usedMove, DungeonPokemon target, Floor floor, boolean missed, ArrayList<DungeonEvent> events);
 
 }
