+++ PMDMMO-common/src/com/darkxell/common/move/MoveRegistry.java
@@ -35,17 +35,10 @@ public final class MoveRegistry
 
 		Element root = XMLUtils.read(MoveRegistry.class.getResourceAsStream("/data/moves.xml"));
 		for (Element e : root.getChildren())
-			try
-			{
-				String className = e.getAttributeValue("movetype");
-				if (className == null) className = "";
-				Class<?> c = Class.forName(Move.class.getName() + className);
-				Move move = (Move) c.getConstructor(Element.class).newInstance(e);
-				moves.put(move.id, move);
-			} catch (Exception e1)
-			{
-				e1.printStackTrace();
-			}
+		{
+			Move move = new Move(e);
+			moves.put(move.id, move);
+		}
 
 		ATTACK = find(0);
 	}
